refactor(slidy-docs): manage kit update breaks
