+++ www/package.json
@@ -18,8 +18,8 @@
         "@slidy/animation": "workspace:*",
         "@slidy/media": "workspace:*",
         "@slidy/svelte": "workspace:*",
-        "@sveltejs/adapter-static": "^1.0.0-next.39",
-        "@sveltejs/kit": "^1.0.0-next.405",
+        "@sveltejs/adapter-static": "1.0.0-next.39",
+        "@sveltejs/kit": "1.0.0-next.415",
         "eslint-config-prettier": "^8.5.0",
         "eslint-plugin-svelte3": "^3.4.1",
         "mdast-util-to-string": "^3.1.0",
