+++ www/src/routes/playground/+page.svelte
@@ -1,24 +1,8 @@
-<!-- <script lang="ts" context="module">
-	import { media, type Queries } from '$lib/media';
-	
-	export async function load({ session }) {
-		let { theme = JSON.stringify(media.matches) } = session.user;
-
-		return {
-			props: {
-				theme: JSON.parse(theme)
-			}
-		};
-	}
-</script> -->
-
 <script lang="ts">
 	import { getRandomSlides } from "@stores/slides";
 	import { Slidy } from "@slidy/svelte";
 	import { stairs } from "@slidy/animation";
 	import "@slidy/svelte/slidy.css";
-	
-	//export let theme: Queries = media.matches;
 </script>
 
 <svelte:head>
